requirements.txt
encar==0.1.4 
flask==2.3.3 
requests==2.31.0 
python-dotenv==1.0.0 
